refactor(plugins): split PluginLoaderService along its jobs
The loader had grown into six jobs in one 1,065-line class: boot
scan/registry, enable/disable lifecycle, uninstall, sandbox host
lifecycle, ingress dispatch, and host-service resolution. The last three
already live in PluginSandboxBridge, PluginCapabilityContext and
PluginHostServices; this finishes the decomposition for the rest.

PluginPackageScanner owns the boot scan — package discovery in the
configured and legacy directories, manifest validation, interrupted-update
recovery, ghost pruning, and registry-entry reconciliation. PluginLifecycle
owns the runtime — enable/disable/unload, the enable concurrency lock,
config and per-session activation writes, and built-in registration.
PluginUninstaller owns removal. The shared path/containment helpers move
to plugin-paths.ts and stay re-exported from the loader module, so every
existing import keeps working.

PluginLoaderService remains the public facade every consumer calls, with
its constructor signature, registry Maps (shared by reference with the
collaborators and the specs that poke them), and the protected
createSandboxHost seam unchanged. Behavior is byte-identical: same events,
log fields and messages. The scanner and uninstaller specs construct the
collaborators directly; their assertions are unchanged from the
loader-level tests they replace.
